LPD-47993 Extract visit method
