+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page_structure/components/StructureTreeContent.js
@@ -9,14 +9,7 @@ import {useEventListener} from '@liferay/frontend-js-react-web';
 import classNames from 'classnames';
 import React, {useCallback, useEffect, useMemo, useRef, useState} from 'react';
 
-import getAllEditables from '../../../../../app/components/fragment_content/getAllEditables';
 import {fromControlsId} from '../../../../../app/components/layout_data_items/Collection';
-import getAllPortals from '../../../../../app/components/layout_data_items/getAllPortals';
-import {EDITABLE_FRAGMENT_ENTRY_PROCESSOR} from '../../../../../app/config/constants/editableFragmentEntryProcessor';
-import {EDITABLE_TYPE_LABELS} from '../../../../../app/config/constants/editableTypeLabels';
-import {EDITABLE_TYPES} from '../../../../../app/config/constants/editableTypes';
-import {FRAGMENT_ENTRY_TYPES} from '../../../../../app/config/constants/fragmentEntryTypes';
-import {FREEMARKER_FRAGMENT_ENTRY_PROCESSOR} from '../../../../../app/config/constants/freemarkerFragmentEntryProcessor';
 import {ITEM_ACTIVATION_ORIGINS} from '../../../../../app/config/constants/itemActivationOrigins';
 import {ITEM_TYPES} from '../../../../../app/config/constants/itemTypes';
 import {
@@ -47,48 +40,13 @@ import {
 import selectCanUpdateEditables from '../../../../../app/selectors/selectCanUpdateEditables';
 import selectCanUpdateItemConfiguration from '../../../../../app/selectors/selectCanUpdateItemConfiguration';
 import selectCanUpdatePageStructure from '../../../../../app/selectors/selectCanUpdatePageStructure';
-import selectLayoutDataItemLabel from '../../../../../app/selectors/selectLayoutDataItemLabel';
-import canActivateEditable from '../../../../../app/utils/canActivateEditable';
 import {DragAndDropContextProvider} from '../../../../../app/utils/drag_and_drop/useDragAndDrop';
-import isMapped from '../../../../../app/utils/editable_value/isMapped';
-import isMappedToCollection from '../../../../../app/utils/editable_value/isMappedToCollection';
-import findPageContent from '../../../../../app/utils/findPageContent';
-import {formIsMapped} from '../../../../../app/utils/formIsMapped';
-import {formIsRestricted} from '../../../../../app/utils/formIsRestricted';
 import getFirstControlsId from '../../../../../app/utils/getFirstControlsId';
-import getMappingFieldsKey from '../../../../../app/utils/getMappingFieldsKey';
-import getSelectedField from '../../../../../app/utils/getSelectedField';
-import isHideable from '../../../../../app/utils/isHideable';
-import {isItemHidden} from '../../../../../app/utils/isItemHidden';
-import isRemovable from '../../../../../app/utils/isRemovable';
 import usePageContents from '../../../../../app/utils/usePageContents';
 import StructureTreeNode from './StructureTreeNode';
 import StructureTreeNodeActions from './StructureTreeNodeActions';
 import VisibilityButton from './VisibilityButton';
-
-const EDITABLE_TYPE_ICONS = {
-	[EDITABLE_TYPES.action]: 'cursor',
-	[EDITABLE_TYPES.backgroundImage]: 'picture',
-	[EDITABLE_TYPES.html]: 'code',
-	[EDITABLE_TYPES.image]: 'picture',
-	[EDITABLE_TYPES.link]: 'link',
-	[EDITABLE_TYPES['rich-text']]: 'text-editor',
-	[EDITABLE_TYPES.text]: 'text',
-};
-
-const LAYOUT_DATA_ITEM_TYPE_ICONS = {
-	[LAYOUT_DATA_ITEM_TYPES.collection]: 'list',
-	[LAYOUT_DATA_ITEM_TYPES.collectionItem]: 'document',
-	[LAYOUT_DATA_ITEM_TYPES.container]: 'container',
-	[LAYOUT_DATA_ITEM_TYPES.form]: 'container',
-	[LAYOUT_DATA_ITEM_TYPES.formStep]: 'arrow-end',
-	[LAYOUT_DATA_ITEM_TYPES.formStepContainer]: 'table',
-	[LAYOUT_DATA_ITEM_TYPES.dropZone]: 'box-container',
-	[LAYOUT_DATA_ITEM_TYPES.fragment]: 'code',
-	[LAYOUT_DATA_ITEM_TYPES.fragmentDropZone]: 'box-container',
-	[LAYOUT_DATA_ITEM_TYPES.root]: 'page',
-	[LAYOUT_DATA_ITEM_TYPES.row]: 'table',
-};
+import visit from './visit';
 
 export default function StructureTreeContent({expandedKeys, setExpandedKeys}) {
 	const activeItemIds = useActiveItemIds();
@@ -399,374 +357,4 @@ export default function StructureTreeContent({expandedKeys, setExpandedKeys}) {
 			</DragAndDropContextProvider>
 		</div>
 	);
-}
-
-function getCollectionAncestor(layoutData, itemId) {
-	const item = layoutData.items[itemId];
-
-	const parent = layoutData.items[item.parentId];
-
-	if (!parent) {
-		return null;
-	}
-
-	return parent.type === LAYOUT_DATA_ITEM_TYPES.collection
-		? parent
-		: getCollectionAncestor(layoutData, item.parentId);
-}
-
-function getDocumentFragment(content) {
-	const fragment = document.createDocumentFragment();
-	const div = document.createElement('div');
-
-	div.innerHTML = content;
-
-	return fragment.appendChild(div);
-}
-
-function getKey({collectionConfig, editable, infoItem, selectedMappingTypes}) {
-	if (collectionConfig) {
-		if (collectionConfig.classNameId) {
-			return getMappingFieldsKey(collectionConfig);
-		}
-		else {
-			return collectionConfig.key;
-		}
-	}
-	else if (editable.mappedField) {
-		return getMappingFieldsKey(selectedMappingTypes);
-	}
-	else if (!infoItem) {
-		return null;
-	}
-
-	return getMappingFieldsKey(infoItem);
-}
-
-function getMappedFieldLabel(
-	editable,
-	collectionConfig,
-	pageContents,
-	mappingFields
-) {
-	const infoItem = findPageContent(pageContents, editable);
-
-	const {selectedMappingTypes} = config;
-
-	if (!infoItem && !selectedMappingTypes && !collectionConfig) {
-		for (const [mappingFieldsKey, fields] of Object.entries(
-			mappingFields
-		)) {
-			if (mappingFieldsKey.startsWith(editable.classNameId)) {
-				const field = getSelectedField({
-					fields,
-					mappingFieldsKey,
-					value:
-						editable.mappedField ||
-						editable.fieldId ||
-						editable.collectionFieldId,
-				});
-
-				return field?.label;
-			}
-		}
-
-		return null;
-	}
-
-	const key = getKey({
-		collectionConfig,
-		editable,
-		infoItem,
-		selectedMappingTypes,
-	});
-	const fields = mappingFields[key];
-
-	if (fields) {
-		const field = getSelectedField({
-			fields,
-			mappingFieldsKey: key,
-			value:
-				editable.mappedField ||
-				editable.fieldId ||
-				editable.collectionFieldId,
-		});
-
-		return field?.label;
-	}
-
-	return null;
-}
-
-function getNameInfo(item) {
-	if (
-		item.type === LAYOUT_DATA_ITEM_TYPES.container &&
-		item.config.htmlTag !== 'div'
-	) {
-		return item.config.htmlTag;
-	}
-
-	return null;
-}
-
-function fragmentIsMapped(item, fragmentEntryLinks) {
-	if (item.type === LAYOUT_DATA_ITEM_TYPES.form) {
-		return formIsMapped(item);
-	}
-	else if (item.type === LAYOUT_DATA_ITEM_TYPES.fragment) {
-		const {editableValues, fragmentEntryType} =
-			fragmentEntryLinks[item.config.fragmentEntryLinkId];
-
-		return fragmentEntryType === FRAGMENT_ENTRY_TYPES.input
-			? Boolean(
-					editableValues[FREEMARKER_FRAGMENT_ENTRY_PROCESSOR]
-						?.inputFieldId
-				)
-			: false;
-	}
-
-	return false;
-}
-
-function visit(
-	item,
-	items,
-	{
-		activeItemIds,
-		canUpdateEditables,
-		canUpdateItemConfiguration,
-		editingNodeId,
-		fragmentEntryLinks,
-		hasHiddenAncestor,
-		hoveredItemId,
-		isMasterPage,
-		layoutData,
-		layoutDataRef,
-		mappingFields,
-		masterLayoutData,
-		onHoverNode,
-		pageContents,
-		restrictedItemIds,
-		selectedViewportSize,
-	}
-) {
-	const children = [];
-
-	const itemInMasterLayout =
-		masterLayoutData &&
-		Object.keys(masterLayoutData.items).includes(item.itemId);
-
-	const hidden = isItemHidden(layoutData, item.itemId, selectedViewportSize);
-
-	let icon = LAYOUT_DATA_ITEM_TYPE_ICONS[item.type];
-
-	if (item.type === LAYOUT_DATA_ITEM_TYPES.fragment) {
-		const fragmentEntryLink =
-			fragmentEntryLinks[item.config.fragmentEntryLinkId];
-
-		icon = fragmentEntryLink.icon || icon;
-
-		const documentFragment = getDocumentFragment(fragmentEntryLink.content);
-
-		const sortedElements = [
-			...getAllEditables(documentFragment),
-			...getAllPortals(documentFragment),
-		].sort((a, b) => a.priority - b.priority);
-
-		const editableTypes = fragmentEntryLink.editableTypes;
-
-		let collectionAncestor = null;
-
-		sortedElements.forEach((element) => {
-			if (element.editableId) {
-				const {editableId} = element;
-
-				const editable =
-					fragmentEntryLink.editableValues[
-						EDITABLE_FRAGMENT_ENTRY_PROCESSOR
-					]?.[editableId];
-
-				const childId = `${item.config.fragmentEntryLinkId}-${editableId}`;
-				const type =
-					editableTypes[editableId] || EDITABLE_TYPES.backgroundImage;
-
-				if (!collectionAncestor) {
-					collectionAncestor = isMappedToCollection(editable)
-						? getCollectionAncestor(
-								fragmentEntryLink.masterLayout
-									? masterLayoutData
-									: layoutData,
-								item.itemId
-							)
-						: null;
-				}
-
-				const collectionConfig = collectionAncestor?.config?.collection;
-
-				const mappedFieldLabel = isMapped(editable)
-					? getMappedFieldLabel(
-							editable,
-							collectionConfig,
-							pageContents,
-							mappingFields
-						)
-					: null;
-
-				children.push({
-					activable:
-						canUpdateEditables &&
-						canActivateEditable(selectedViewportSize, type),
-					active: activeItemIds.includes(childId),
-					children: [],
-					draggable: false,
-					hidden: false,
-					hiddenAncestor: hasHiddenAncestor || hidden,
-					hideable: false,
-					hovered: childId === hoveredItemId,
-					icon: EDITABLE_TYPE_ICONS[type],
-					id: childId,
-					isMasterItem: !isMasterPage && itemInMasterLayout,
-					itemType: ITEM_TYPES.editable,
-					mapped: isMapped(editable),
-					name: mappedFieldLabel || editableId,
-					onHoverNode,
-					parentId: item.itemId,
-					removable: false,
-					tooltipTitle: EDITABLE_TYPE_LABELS[type],
-				});
-			}
-			else {
-				const {dropZoneId, mainItemId} = element;
-
-				children.push({
-					...visit(items[mainItemId], items, {
-						activeItemIds,
-						canUpdateEditables,
-						canUpdateItemConfiguration,
-						editingNodeId,
-						fragmentEntryLinks,
-						hasHiddenAncestor: hasHiddenAncestor || hidden,
-						hoveredItemId,
-						isMasterPage,
-						layoutData,
-						layoutDataRef,
-						mappingFields,
-						masterLayoutData,
-						onHoverNode,
-						pageContents,
-						restrictedItemIds,
-						selectedViewportSize,
-					}),
-
-					name: `${Liferay.Language.get('drop-zone')} ${dropZoneId}`,
-					removable: false,
-				});
-			}
-		});
-	}
-	else {
-		item.children.forEach((childId) => {
-			if (
-				(item.type === LAYOUT_DATA_ITEM_TYPES.collection &&
-					(!item.config.collection ||
-						restrictedItemIds.has(item.itemId))) ||
-				(item.type === LAYOUT_DATA_ITEM_TYPES.form &&
-					(!formIsMapped(item) || formIsRestricted(item)))
-			) {
-				return;
-			}
-
-			const childItem = items[childId];
-
-			if (
-				!isMasterPage &&
-				childItem.type === LAYOUT_DATA_ITEM_TYPES.dropZone
-			) {
-				const dropZoneChildren = visit(
-					layoutData.items[layoutData.rootItems.main],
-					layoutData.items,
-					{
-						activeItemIds,
-						canUpdateEditables,
-						canUpdateItemConfiguration,
-						editingNodeId,
-						fragmentEntryLinks,
-						hasHiddenAncestor: hasHiddenAncestor || hidden,
-						hoveredItemId,
-						isMasterPage,
-						layoutData,
-						layoutDataRef,
-						mappingFields,
-						masterLayoutData,
-						onHoverNode,
-						pageContents,
-						restrictedItemIds,
-						selectedViewportSize,
-					}
-				).children;
-
-				children.push(...dropZoneChildren);
-			}
-			else {
-				const child = visit(childItem, items, {
-					activeItemIds,
-					canUpdateEditables,
-					canUpdateItemConfiguration,
-					editingNodeId,
-					fragmentEntryLinks,
-					hasHiddenAncestor: hasHiddenAncestor || hidden,
-					hoveredItemId,
-					isMasterPage,
-					layoutData,
-					layoutDataRef,
-					mappingFields,
-					masterLayoutData,
-					onHoverNode,
-					pageContents,
-					restrictedItemIds,
-					selectedViewportSize,
-				});
-
-				children.push(child);
-			}
-		});
-	}
-
-	return {
-		activable:
-			(Liferay.FeatureFlags['LPD-18221'] ||
-				(item.type !== LAYOUT_DATA_ITEM_TYPES.formStep &&
-					item.type !== LAYOUT_DATA_ITEM_TYPES.column &&
-					item.type !== LAYOUT_DATA_ITEM_TYPES.fragmentDropZone)) &&
-			item.type !== LAYOUT_DATA_ITEM_TYPES.collectionItem &&
-			canUpdateItemConfiguration,
-		active: activeItemIds.includes(item.itemId),
-		children,
-		config: layoutDataRef?.current?.items[item.itemId]?.config,
-		draggable: true,
-		hidden,
-		hiddenAncestor: hasHiddenAncestor,
-		hideable:
-			!itemInMasterLayout &&
-			isHideable(item, fragmentEntryLinks, layoutData),
-		hovered: item.itemId === hoveredItemId,
-		icon,
-		id: item.itemId,
-		isMasterItem: !isMasterPage && itemInMasterLayout,
-		itemType: ITEM_TYPES.layoutDataItem,
-		mapped: fragmentIsMapped(item, fragmentEntryLinks),
-		name: selectLayoutDataItemLabel({fragmentEntryLinks, layoutData}, item),
-		nameInfo: getNameInfo(item),
-		onHoverNode,
-		parentItemId: item.parentId,
-		removable: !itemInMasterLayout && isRemovable(item, layoutData),
-		tooltipTitle: selectLayoutDataItemLabel(
-			{fragmentEntryLinks, layoutData},
-			item,
-			{
-				useCustomName: false,
-			}
-		),
-		type: item.type,
-	};
 }
